+++ src/main/java/miyucomics/hexical/mixin/FabricIXplatImplMixin.java
@@ -1,19 +1,63 @@
 package miyucomics.hexical.mixin;
 
+import at.petrak.hexcasting.api.addldata.ADHexHolder;
 import at.petrak.hexcasting.api.addldata.ADIotaHolder;
+import at.petrak.hexcasting.api.casting.iota.Iota;
+import at.petrak.hexcasting.api.pigment.FrozenPigment;
 import at.petrak.hexcasting.fabric.xplat.FabricXplatImpl;
 import miyucomics.hexical.entities.AnimatedScrollEntity;
+import miyucomics.hexical.utils.CharmedItemUtilities;
 import net.minecraft.entity.Entity;
+import net.minecraft.item.ItemStack;
+import net.minecraft.server.world.ServerWorld;
+import org.jetbrains.annotations.Nullable;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
-@Mixin(FabricXplatImpl.class)
+import java.util.List;
+
+@Mixin(value = FabricXplatImpl.class, remap = false)
 public class FabricIXplatImplMixin {
 	@Inject(method = "findDataHolder(Lnet/minecraft/entity/Entity;)Lat/petrak/hexcasting/api/addldata/ADIotaHolder;", at = @At("HEAD"), cancellable = true)
 	private void readFromLivingScrolls(Entity entity, CallbackInfoReturnable<ADIotaHolder> cir) {
 		if (entity instanceof AnimatedScrollEntity)
 			cir.setReturnValue((AnimatedScrollEntity) entity);
+	}
+
+	@Inject(method = "findHexHolder", at = @At("HEAD"), cancellable = true)
+	public void makeCharmedItemsProvideHexHolder(ItemStack stack, CallbackInfoReturnable<ADHexHolder> cir) {
+		if (CharmedItemUtilities.isStackCharmed(stack)) {
+			cir.setReturnValue(new ADHexHolder() {
+				@Override
+				public boolean canDrawMediaFromInventory() {
+					return false;
+				}
+
+				@Override
+				public boolean hasHex() {
+					return true;
+				}
+
+				@Override
+				public @Nullable List<Iota> getHex(ServerWorld level) {
+					return CharmedItemUtilities.getHex(stack, level);
+				}
+
+				@Override
+				public void writeHex(List<Iota> patterns, @Nullable FrozenPigment pigment, long media) { }
+
+				@Override
+				public void clearHex() {
+					CharmedItemUtilities.removeCharm(stack);
+				}
+
+				@Override
+				public @Nullable FrozenPigment getPigment() {
+					return null;
+				}
+			});
+		}
 	}
 }
